Python Pokemon Game Engine
This is a Pokemon game engine written in python with a goal of no
library dependencies other than the standard library.

Right now the goal is create a full text based battle game that mimics
Pokemon with an easy to import Pokemon generator.

Long-term goals for the project are a full release as well as support
for servers to battle together.

Features Currently:
- Rudimentary Enemy AI
-Attacks
-Class system for Pokemon, Trainers, Moves, and Items

Planned Features
-Robust AI stronger than commercial game
-Import of first generation pokemon and movesets
-Stat changes, status changes, etc
-One on one Mode
-Robust UI